==> notebooks/3_Modelagem.ipynb <==
## Bibliotecas

import pandas as pd
import numpy as np
import joblib 
from sklearn.linear_model import LinearRegression
from sklearn.metrics import mean_squared_error, r2_score
from sklearn.ensemble import RandomForestRegressor
import xgboost as xgb

## 1. Carregar os dados salvos do pre_processamento.

X_train_processado = joblib.load('../data_base/processed/X_train_processed.pkl')
X_test_processado = joblib.load('../data_base/processed/X_test_processed.pkl')
y_train = joblib.load('../data_base/processed/y_train.pkl')
y_test = joblib.load('../data_base/processed/y_test.pkl')

## 2. Modelo 1: Regressão Linear

Utilizando os testes e treinos processados.

#Criando o modelo com LinearRegression
modelo_linear = LinearRegression()

#Treinando o modelo com fit
modelo_linear.fit(X_train_processado, y_train)

#Previsões do modelo linear com predict
previsoes_linear = modelo_linear.predict(X_test_processado)
print('Modelo treinado')


### 2.2 Avaliando o modelo com metricas e erro padrao.

#Calculando R²
r2_linear = r2_score(y_test, previsoes_linear)
print(f'R2 quadrado do modelo linear: {r2_linear:.4f}')
print(f"(Isso significa que o modelo explica {r2_linear*100:.2f}% da variação dos preços)")

print('\n')

#Calculando o RMSE
rmse_linear = np.sqrt(mean_squared_error(y_test, previsoes_linear))
print(f'RMSE (ERRO MEDIO) do modelo Linear: {rmse_linear:.2f}')
print(f"(Isso significa que, em média, o modelo erra a previsão do preço em R$ {rmse_linear:.2f})")

## 3. Modelo 2: Random Forest Regressor
Obs: Na regressão linear tivemos um erro padrao muito alto, vamos observa esse.

#Criando o modelo
modelo_rf = RandomForestRegressor(random_state=42, n_jobs=-1)

#Treinando o modelo
print('Treinando o Random Forest: ')
modelo_rf.fit(X_train_processado, y_train)

#Fazendo a previsção
previsoes_rf = modelo_rf.predict(X_test_processado)

print('Modelo Random Forest treinado')

### 3.1 Avaliação do modelo.

#Calculando o r2 do Random Forest
r2_rf  = r2_score(y_test, previsoes_rf)
print(f'R2 do Random Forest: {r2_rf:.4f} é {r2_rf*100:.2f}%')

print('\n')

#Calcular o RMSE do Randow Forest
rmse_rf = np.sqrt(mean_squared_error(y_test, previsoes_rf))
print(f'RMSE do modelo Random Forest: {rmse_rf:.2f}')

### 3.2 Comparando o RMSE dos dois modelos.


#RMSE Linear
print(f'RMSE do modelo Linear:        {rmse_linear:.2f}')
print(f'RMSE do modelo Random Forest: {rmse_rf:.2f}')

## 4. Modelo 3: XGBoost Regressor
Obs: O randow forest teve um desempenho excelente, vamos testar se o XGBoost supera.


#Criando o modelo xgb
modelo_xgb  = xgb.XGBRegressor(random_state=42, n_jobs=-1)

#Treinar o modelo
print('Treinando o modelo:')
modelo_xgb.fit(X_train_processado, y_train)

#Previsão do modleo
previsoes_xgb = modelo_xgb.predict(X_test_processado)

print('Modelo xbg treinado: ')

### 4.1 Avaliando o modelo.

#R2 do modelo.
r2_xgb = r2_score(y_test, previsoes_xgb)
print(f'R2 do modelo xgb: {r2_xgb:.4f} isso é {r2_xgb*100:.2f}%')

#RMSE do modelo xgb.
rmse_xgb = np.sqrt(mean_squared_error(y_test, previsoes_xgb))
print(f'RMSE do modelo xgb: {rmse_xgb:.2f}')

## 5 Comparando o 3 modelos.

#Modelo Linear
print(f'RMSE do modelo Linear:        {rmse_linear:.2f}')

#Modelo Randow Forest
print(f'RMSE do modelo Random Forest: {rmse_rf:.2f}')

#Modelo XGB
print(f'RMSE do modelo xgb:           {rmse_xgb:.2f}')
